review comment: mv counters struct to comments.go
